Control steps from Int to Double. Define tstep_refinement_ in Settings/Model class, transfer tstep_refinement_ from parent class.

+++ FieldOpt/Simulation/simulator_interfaces/driver_file_writers/driver_parts/ecl_driver_parts/schedule_section.cpp
@@ -34,7 +34,7 @@ namespace ECLDriverParts {
 using Printer::num2str;
 
 Schedule::Schedule(QList<Model::Wells::Well *> *wells,
-                   const QList<int>& control_times,
+                   const QList<double>& control_times,
                    const QList<int>& start_date,
                    ScheduleInsets &insets,
                    Settings::Settings *settings) : ECLDriverPart(settings) {
